Fix to deploy

+++ src/pages/Barang/DataBarang.js
@@ -4,7 +4,6 @@ import Card from "../../components/elements/Card";
 import { useHistory } from "react-router-dom";
 import styled from "styled-components";
 import Modal from "../../components/elements/Modal";
-import { Button, ButtonLight } from "../../components/elements/Button";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
 import { Helmet } from "react-helmet";
@@ -78,6 +77,7 @@ const DataBarang = () => {
 
   const fetchBarang = async () => {
     setShowLoading(true);
+
     await api
       .get("/barang", {
         params: {
@@ -106,7 +106,7 @@ const DataBarang = () => {
               harga_beli: value.harga_beli,
               kategori: value.id_kategori.nama_kategori,
               satuan: value.id_satuan.nama_satuan,
-            })
+            });
           });
           return data;
         });
@@ -125,25 +125,19 @@ const DataBarang = () => {
       .catch((error) => {
         // Unauthorized
         console.log(error);
-        if (error.response.status === 401) {
+        if (error.response && error.response.status === 401) {
           localStorage.clear();
           return history.push("/login");
         }
 
-        setAlert({
-          ...alert,
-          status: 500,
-          message: "Internal server error!",
-          error: true,
-        });
+        setAlert({ message: "Internal server error!", error: true });
         setShowAlert(true);
       });
+
     setShowLoading(false);
   };
 
-  const handleSubmitDeleteBarang = async (e) => {
-    e.preventDefault();
-    setShowAlert(false);
+  const handleSubmitDeleteBarang = async () => {
     setShowLoading(true);
 
     await api
@@ -154,26 +148,22 @@ const DataBarang = () => {
       })
       .then((response) => {
         fetchBarang();
-        setAlert({ ...alert, ...response.data });
+        setAlert({ message: response.data.message, error: false });
         setShowAlert(true);
         setShowModalDeleteBarang(false);
         setFormDataDeleteBarang(initialStateFormDataDeleteBarang); // reset form
       })
       .catch((error) => {
         // Unauthorized
-        if (error.response.status === 401) {
+        if (error.response && error.response.status === 401) {
           localStorage.clear();
           return history.push("/login");
         }
 
-        setAlert({
-          ...alert,
-          status: 500,
-          message: "Internal server error!",
-          error: true,
-        });
+        setAlert({ message: "Internal server error!", error: true });
         setShowAlert(true);
       });
+
     setShowLoading(false);
   };
 
@@ -211,18 +201,16 @@ const DataBarang = () => {
           </td>
           <td className="border">
             <div className="flex items-center justify-center text-xs space-x-1">
-              <ButtonLight
-                variant="pill"
+              <button
+                className="border border-indigo-300 bg-indigo-50 hover:bg-indigo-200 text-indigo-600 rounded-full focus:ring focus:ring-indigo-100 focus:outline-none px-4 py-1.5"
                 onClick={() => {
                   history.push(`/barang/${value._id}/edit`);
                 }}>
                 Edit
-              </ButtonLight>
-              <ButtonLight
-                theme="red"
-                variant="pill"
+              </button>
+              <button
+                className="border border-red-300 bg-red-50 hover:bg-red-200 text-red-600 rounded-full focus:ring focus:ring-red-100 focus:outline-none px-4 py-1.5"
                 onClick={() => {
-                  console.log('click hapus')
                   setShowModalDeleteBarang(true);
                   setFormDataDeleteBarang({
                     _id: value._id,
@@ -230,7 +218,7 @@ const DataBarang = () => {
                   });
                 }}>
                 Hapus
-              </ButtonLight>
+              </button>
             </div>
           </td>
         </Tr>
@@ -244,30 +232,47 @@ const DataBarang = () => {
 
   return (
     <>
-      <Helmet><title>Data Barang | INVENTORY</title></Helmet>
+      <Helmet>
+        <title>Data Barang | INVENTORY</title>
+      </Helmet>
       {showLoading ? (
         <div className="fixed bg-transparent w-full h-full z-30">
           <div
             className="fixed top-1/2 left-1/2 text-white transform -translate-y-1/2 -translate-x-1/2 rounded-lg px-8 py-3"
             style={{ backgroundColor: "#00000097" }}>
             <Loading>
-              <div className="font-montserrat mt-2">loading...</div>
+              <div className="font-montserrat text-gray-300 mt-2">
+                Loading...
+              </div>
             </Loading>
           </div>
         </div>
       ) : null}
       <Alert
         show={showAlert}
-        {...alert}
-        afterClose={() => setShowAlert(false)}
-      />
+        afterClose={() => {
+          setShowAlert(false);
+        }}>
+        {alert.error ? (
+          <div
+            className={`bg-red-300 font-bold text-sm text-white rounded-lg px-8 py-3`}>
+            {alert.message}
+          </div>
+        ) : (
+          <div
+            className={`bg-green-300 font-bold text-sm text-white rounded-lg px-8 py-3`}>
+            {alert.message}
+          </div>
+        )}
+      </Alert>
+      
       <Modal
         show={showModalDeleteBarang}
         afterClose={() => setShowModalDeleteBarang(false)}>
         <Card className="font-montserrat">
           <div className="flex items-start justify-between mb-4">
             <div className="font-bold text-gray-500 text-lg border-b pb-2">
-              Delete Barang
+              Hapus Barang
             </div>
             <button
               onClick={() => {
@@ -279,51 +284,46 @@ const DataBarang = () => {
               />
             </button>
           </div>
-          <form onSubmit={handleSubmitDeleteBarang}>
-            <div className="text-sm">
-              Anda yakin ingin menghapus barang{" "}
-              <strong>{formDataDeleteBarang.nama_barang}</strong>?
-            </div>
-            <div className="flex justify-between text-sm space-x-2 mt-8">
-              <Button
-                type="button"
-                theme="red"
-                onClick={() => {
-                  setShowModalDeleteBarang(false);
-                }}>
-                Batal
-              </Button>
-              <Button type="submit" theme="green">
-                Ya
-              </Button>
-            </div>
-          </form>
+          <div className="text-sm">
+            Anda yakin ingin menghapus barang{" "}
+            <strong>{formDataDeleteBarang.nama_barang}</strong>?
+          </div>
+          <div className="flex justify-between text-sm space-x-2 mt-8">
+            <button
+              className="bg-red-500 hover:bg-red-400 text-red-100 rounded focus:ring focus:ring-red-100 focus:outline-none px-4 py-1.5"
+              onClick={() => {
+                setShowModalDeleteBarang(false);
+              }}>
+              Batal
+            </button>
+            <button
+              className="bg-green-500 hover:bg-green-400 text-green-100 rounded focus:ring focus:ring-green-100 focus:outline-none px-4 py-1.5"
+              onClick={handleSubmitDeleteBarang}>
+              Ya
+            </button>
+          </div>
         </Card>
       </Modal>
 
-      <Card>
-        <div className="font-montserrat font-bold text-gray-500 text-xl mb-6">
-          Data Barang
-        </div>
-        <Button
-          className="mb-4"
+      <Card className="font-montserrat">
+        <div className="font-bold text-gray-500 text-xl mb-6">Data Barang</div>
+        <button
+          className="bg-indigo-500 hover:bg-indigo-400 text-indigo-100 rounded focus:ring focus:ring-indigo-100 focus:outline-none px-4 py-1.5 mr-2 mb-4"
           onClick={() => {
             history.push("/barang/tambah");
           }}>
           Tambah Barang
-        </Button>
+        </button>
 
-        <ButtonLight className="text-sm ml-4">
-          <FontAwesomeIcon icon={faFileAlt} />
-          <CSVLink
-            className="ml-2"
-            headers={headersCSV}
-            data={dataCSV}
-            filename="Data_Barang_INVENTORY.csv"
-            target="_blank">
-            Export
-          </CSVLink>
-        </ButtonLight>
+        <CSVLink
+          className="border border-indigo-300 bg-indigo-50 hover:bg-indigo-200 text-indigo-600 rounded focus:ring focus:ring-indigo-100 focus:outline-none px-4 py-1.5 ml-2"
+          headers={headersCSV}
+          data={dataCSV}
+          filename="Data_Barang_INVENTORY.csv"
+          target="_blank">
+          <FontAwesomeIcon icon={faFileAlt} className="mr-2" />
+          Export
+        </CSVLink>
 
         <Datatable
           page={dataTable.page}
