+++ src/pages/Barang/TambahBarang.js
@@ -1,7 +1,6 @@
-import axios from "axios";
+import api from "../../config/api";
 import { useEffect, useState } from "react";
 import Card from "../../components/elements/Card";
-import { Button } from "../../components/elements/Button";
 import { Helmet } from "react-helmet";
 import Alert from "../../components/elements/Alert";
 import Loading from "../../components/elements/Loading";
@@ -29,43 +28,42 @@ const TambahBarang = () => {
   const [showLoading, setShowLoading] = useState(false);
   const [showAlert, setShowAlert] = useState(false);
   const [alert, setAlert] = useState({
-    status: 200,
     message: "",
-    duration: 3000,
+    error: false,
   });
   const history = useHistory();
 
   const fetchKategori = async () => {
-    await axios
-      .get("/kategori")
+    await api
+      .get("/kategori", {
+        headers: { "x-access-token": localStorage.getItem("token") },
+      })
       .then((response) => {
         setDataKategori(response.data.data);
       })
       .catch((error) => {
         // Unauthorized
-        if (error.response.status === 401) {
+        if (error.response && error.response.status === 401) {
           localStorage.clear();
           return history.push("/login");
         }
-
-        console.log(error);
       });
   };
 
   const fetchSatuan = async () => {
-    await axios
-      .get("/satuan")
+    await api
+      .get("/satuan", {
+        headers: { "x-access-token": localStorage.getItem("token") },
+      })
       .then((response) => {
         setDataSatuan(response.data.data);
       })
       .catch((error) => {
         // Unauthorized
-        if (error.response.status === 401) {
+        if (error.response && error.response.status === 401) {
           localStorage.clear();
           return history.push("/login");
         }
-
-        console.log(error);
       });
   };
 
@@ -75,11 +73,7 @@ const TambahBarang = () => {
     Object.entries(formData).forEach((data) => {
       let [name, value] = data;
 
-      if (name === "stok" || name === "harga_jual" || name === "harga_beli") {
-        value = value.replace(/\D/g, "");
-      }
-
-      if (!value) {
+      if (value === undefined || value === "") {
         isError = true;
         setFormDataError((state) => ({ ...state, [name]: "harus diisi" }));
       }
@@ -95,14 +89,14 @@ const TambahBarang = () => {
     // only number
     if (name === "stok" || name === "harga_jual" || name === "harga_beli") {
       value = value.replace(/\D/g, "");
+      value = value === "" ? '' : parseInt(value);
     }
 
-    if (value) {
-      setFormData((state) => ({ ...state, [name]: value }));
-      setFormDataError((state) => ({ ...state, [name]: false }));
-    } else {
-      setFormData((state) => ({ ...state, [name]: value }));
+    setFormData((state) => ({ ...state, [name]: value }));
+    if (value === undefined || value === "") {
       setFormDataError((state) => ({ ...state, [name]: "harus diisi" }));
+    } else {
+      setFormDataError((state) => ({ ...state, [name]: false }));
     }
   };
 
@@ -112,32 +106,28 @@ const TambahBarang = () => {
     if (!formValidation()) {
       return false;
     }
+
     setShowLoading(true);
 
-    await axios
+    await api
       .post("/barang", formData, {
-        headers: {
-          "Content-Type": "application/json",
-        },
+        headers: { "x-access-token": localStorage.getItem("token") },
       })
       .then((response) => {
-        setAlert((state) => ({ ...state, ...response.data }));
+        setAlert({ message: response.data.message, error: false });
         setShowAlert(true);
       })
       .catch((error) => {
         // Unauthorized
-        if (error.response.status === 401) {
+        if (error.response && error.response.status === 401) {
           localStorage.clear();
           return history.push("/login");
         }
 
-        setAlert((state) => ({
-          ...state,
-          status: 500,
-          message: "Internal server error!",
-        }));
+        setAlert({ message: "Internal server error!", error: true });
         setShowAlert(true);
       });
+
     setShowLoading(false);
   };
 
@@ -158,29 +148,44 @@ const TambahBarang = () => {
       </Helmet>
       {showLoading ? (
         <div className="fixed bg-transparent w-full h-full z-30">
-          <div className="fixed top-1/2 left-1/2 transform -translate-y-1/2 -translate-x-1/2">
+          <div
+            className="fixed top-1/2 left-1/2 text-white transform -translate-y-1/2 -translate-x-1/2 rounded-lg px-8 py-3"
+            style={{ backgroundColor: "#00000097" }}>
             <Loading>
-              <div className="font-montserrat mt-2">loading...</div>
+              <div className="font-montserrat text-gray-300 mt-2">
+                Loading...
+              </div>
             </Loading>
           </div>
         </div>
       ) : null}
       <Alert
         show={showAlert}
-        {...alert}
         afterClose={() => {
           setShowAlert(false);
-          if (alert.status === 200) {
-            history.goBack();
+          if (alert.error === false) {
+            return history.goBack();
           }
-        }}
-      />
+        }}>
+        {alert.error ? (
+          <div
+            className={`bg-red-300 font-bold text-sm text-white rounded-lg px-8 py-3`}>
+            {alert.message}
+          </div>
+        ) : (
+          <div
+            className={`bg-green-300 font-bold text-sm text-white rounded-lg px-8 py-3`}>
+            {alert.message}
+          </div>
+        )}
+      </Alert>
+
       <Card className="font-montserrat w-full sm:w-4/5 md:w-3/4 lg:w-2/3 xl:w-1/2 mx-auto">
         <div className="font-montserrat font-bold text-lg text-gray-500 mb-6">
           Tambah Barang
         </div>
         <form onSubmit={handleSubmit}>
-          <div className="flex flex-col justify-center text-sm space-y-4">
+          <div className="flex flex-col justify-center space-y-4">
             <div className="grid grid-cols-12 items-center gap-x-4 gap-y-1">
               <div className="col-span-full md:col-span-4">
                 Nama Barang <span className="text-red-400">*</span>
@@ -196,7 +201,7 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.nama_barang ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Nama barang ${formDataError.nama_barang}`}
               </div>
             </div>
@@ -215,7 +220,7 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.stok ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Stok ${formDataError.stok}`}
               </div>
             </div>
@@ -237,7 +242,7 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.harga_jual ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Harga jual ${formDataError.harga_jual}`}
               </div>
             </div>
@@ -259,7 +264,7 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.harga_beli ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Harga beli ${formDataError.harga_beli}`}
               </div>
             </div>
@@ -268,7 +273,7 @@ const TambahBarang = () => {
                 Kategori <span className="text-red-400">*</span>
               </div>
               <select
-                className="col-span-full md:col-span-8 border border-gray-300 rounded-md focus:ring focus:ring-indigo-200 focus:outline-none p-2"
+                className="col-span-full md:col-span-8 bg-white border border-gray-300 rounded-md focus:ring focus:ring-indigo-200 focus:outline-none p-2"
                 value={formData.id_kategori}
                 name="id_kategori"
                 onChange={handleChange}>
@@ -286,7 +291,7 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.id_kategori ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Kategori ${formDataError.id_kategori}`}
               </div>
             </div>
@@ -295,7 +300,7 @@ const TambahBarang = () => {
                 Satuan <span className="text-red-400">*</span>
               </div>
               <select
-                className="col-span-full md:col-span-8 border border-gray-300 rounded-md focus:ring focus:ring-indigo-200 focus:outline-none p-2"
+                className="col-span-full md:col-span-8 bg-white border border-gray-300 rounded-md focus:ring focus:ring-indigo-200 focus:outline-none p-2"
                 value={formData.id_satuan}
                 name="id_satuan"
                 onChange={handleChange}>
@@ -313,14 +318,14 @@ const TambahBarang = () => {
               <div
                 className={`${
                   formDataError.id_satuan ? "" : "hidden"
-                } md:col-start-5 col-span-full text-xs text-red-400`}>
+                } md:col-start-5 col-span-full text-sm text-red-400`}>
                 {`Satuan ${formDataError.id_satuan}`}
               </div>
             </div>
           </div>
-          <div className="flex justify-end mt-6">
-            <Button className="text-sm">Simpan</Button>
-          </div>
+          <button className="bg-indigo-500 hover:bg-indigo-400 text-indigo-100 rounded focus:ring focus:ring-indigo-100 focus:outline-none w-full px-4 py-1.5 mt-6">
+            Simpan
+          </button>
         </form>
       </Card>
     </>
